Import develop change

+++ app/controllers/managements_controller.rb
@@ -67,6 +67,9 @@ class ManagementsController < ApplicationController
   # @params [Integer] excluded_node_id
   # Create an exclusion between two managements
   def create_exclusion
+    @management = HealthCares::Management.new
+    authorize @management
+
     @management_exclusion = NodeExclusion.new(management_exclusion_params)
     @management_exclusion.node_type = :management
     if @management_exclusion.save
@@ -81,6 +84,9 @@ class ManagementsController < ApplicationController
   # @params [Integer] excluded_node_id
   # Remove an exclusion between two managements
   def remove_exclusion
+    @management = HealthCares::Management.new
+    authorize @management
+
     @management_exclusion = NodeExclusion.management.find_by(management_exclusion_params)
     if @management_exclusion.destroy
       redirect_to algorithm_url(@algorithm, panel: 'managements_exclusions'), notice: t('flash_message.success_updated')
